net: ethernet: xilinx: Fix issues with the optional phy_type property
commit  f5d79fc442335a1a79bbed441d0130e680779162 from
https://github.com/Xilinx/linux-xlnx.git

The xlnx,phy-type is a optional DT property.
If it is not specified in the device tree then it defaults to 0
Which is the value for XAE_PHY_TYPE_MII which then causes an error
When trying to switch from 10/100 to 1000 when in SGMII mode.

Reported-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>