drm: xilinx: dp: Skip the aux communication when there's no sink
commit  1c776fdf104db40d703a54e43f82c11cf1881e46 from
https://github.com/Xilinx/linux-xlnx.git

The DRM master may request the aux transaction, for example, power off
request, even when the sink is not connected. This would result in
the timeout. The connection status is stored in the HPD handler
and checked before the aux transaction.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>