MLK-16694-2: ARM64: dts: fix assigned-clocks for audio device node
commit  3a2df9aab44d68c78ee82d4a8c272ce57d8a5759 from
https://source.codeaurora.org/external/imx/linux-imx.git

Even the clock is not used by current device, but it is used by
other devices, it also need to be included in the assigned-clocks
list. For in kernel side, clock rate is stored, but in scfw
the clock rate is cleared when power off, this mismatch cause
the parent rate is not set in next device, then children clock rate
is wrong.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>