MLK-16709 dts: imx8qm: Disable the UART2 on MEK base board
commit  55e167d1392f10558ca18f436368ce066aeaac58 from
https://source.codeaurora.org/external/imx/linux-imx.git

CM4_1 core will use the UART2 on QM MEK base board as its console. Since this port currently
is a backup debug port on A core side, not really used. We disable it in dts to yield the port
for CM4_1.

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>