ARM: OMAP2: Add functions to save and restore omap hwmod context en-masse.
This is used to support suspend modes like RTC-only and hibernate where
the state of these registers is lost.

After the PRCM loses context in the case of an RTC+DDR cycle omap_hwmod
attempts to return all hwmods to their previous state, however certain
hwmods cannot just be disabled when in their default state, which is why
they need the special handling present in that patch when no driver is
present.

In RTC+DDR mode, even if all drivers are present, the modules are all
returned to their previous state before any driver resume happens so we
will still face the issue described above. This can be prevented by
calling _reidle on all hwmods that need it for any module that is being
disabled to return to it's previous state.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
[d-gerlach@ti.com] Added _reidle part and gfx hwmod fix
Signed-off-by: Russ Dill <Russ.Dill@ti.com>
[j-keerthy@ti.com] squashed gfx hwmod fix ported to 4.14
Signed-off-by: Keerthy <j-keerthy@ti.com>

(cherry picked from commit 87bddafceec754440820ba63ea9c8fa1080b6392)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>