soc: ti: wkup_m3_ipc: Fix misassigned struct device pointer
The firmware loading callback for the volt scaling firmware expects the
the second argument to be a pointer to a struct wkup_m3_ipc, however the
function also attempts to cast it to be used directly as a struct device
pointer. This incorrect pointer is only used for dev_err messages so if
a failure occurs it will result in a crash.

Instead let's use the struct device pointer inside the struct
wkup_m3_ipc pointer that is actually passed so that we have a valid
dev pointer.

Fixes: 1be271922084 ("soc: ti: wkup_m3_ipc: Add support for i2c voltage
scaling")
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Acked-by: Suman Anna <s-anna@ti.com>

(cherry picked from commit f2c1db89f83e3de64f8e539fd66d5c23871c1cc2)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>