mmc: renesas_sdhi: Set the SCC register address immediately
commit e8ec7e0ae74ba7481a61fb3d7830bbcc59c2eb44 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

By setting the address of the SCC register immediately
after starting the kernel, the processing that was returning
at the beginning of reset_hs400_mode is unnecessary.
As a result, reset_hs400_mode is processed even when the
kernel is started in HS400 mode.

Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>
[saito: rebased on v4.14.57, struct renesas_sdhi was moved to renesas_sdhi.h]
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>