mmc: tmio: Add eMMC HS400 mode support
commit e4dd60fa4ee30b7a577cfffc37d10f08a9f258b2 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

This patch adds processing for selecting HS400 mode.

The number of TAPs in HS400 mode differs according to SoC.
The selection range of the sampling clock is 4 or 8,
and in the case of 4 TAP, the SCC error is not checked.

Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>