MLK-17133-01 tty: serial: lpuart: only enable wakeup interrupt when wakeup enabled
commit  c3112f5694da68efb66f0e3aed1fdbbd4fbeb4bd from
https://source.codeaurora.org/external/imx/linux-imx.git

Current driver suppose system disable irq when wakeup is not enabled
like below follow, so it always enable the wakeup interrupt in .suspend_noirq().
	dpm_suspend_noirq()
		device_wakeup_arm_wake_irqs()
			if (device_may_wakeup(wirq->dev))
				enable_irq_wake(wirq->irq);
					irq_set_irq_wake(irq, 1);
		suspend_device_irqs();
			if (irqd_is_wakeup_set(&desc->irq_data))
				__disable_irq(desc);
		device_suspend_noirq(dev);
		...

But in i.MX8x chips, the gic-v3 chip->irq_disable() is not implemented,
so the device's irq line is not masked in noirq stage. Then lpuart interrupt
can wake up system even if it is not enabled as wakeup source.

To avoid the issue, only enable wakeup interrupt when it is enabled as
wakeup source.

Signed-off-by: Fugang Duan <fugang.duan@nxp.com>
Reviewed-by: Robin Gong <yibin.gong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>