MLK-17257-2: drm: imx: dcss: use the WRSCL/RDSRC modules
commit  7a3c40db2860b5584ae74230f2474cd37c9f80e0 from
https://source.codeaurora.org/external/imx/linux-imx.git

This patch makes the necessary changes so that, for downscaling ratios
more than 3:1 and up to 7:1 (for video) and 5:1 (for graphics), the
WRSCL/RDSRC path will be used. This way the DRAM bandwidth will be lower
and spread evenly across the frame time.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>