MLK-17311-1 gpu: imx: dpu: common: Set SC_C_KACHUNK_CNT as 32
commit  6cee049aeab321357c20960855d48ddfebb4f995 from
https://source.codeaurora.org/external/imx/linux-imx.git

The SC_C_KACHUNK_CNT is for dpu blit and represents
how many cycle counts is need to trigger DPR after
DPU shadow being loaded. The initial value is 0x20,
and will change to 0 after the first frame if not set.
So it need be set with a value greater than 0x20.

Signed-off-by: Meng Mingming <mingming.meng@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>