drm/vc4: Fix warning about vblank interrupts before DRM core is ready.
commit f36b84f928b480211a672144b13fe1fd275e79cf from
https://github.com/raspberrypi/linux.git

The SMICS interrupt fires continuously, but since it's 1/100 the rate
of the USB interrupts, we don't really need a way to turn it off.  We
do need to make sure that we don't tell DRM about it until DRM has
asked for the interrupt at least once, because otherwise it will throw
a warning at boot time.

Signed-off-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>