nand: cavium: Nand flash controller for Cavium ARM64 SOCs
Add a driver for the nand flash controller as found on Cavium's
ARM64 SOCs.

The nand flash controller can support up to 8 chips and is presented
as a PCI device. It uses a DMA engine to transfer data between the
nand and L2/DRAM and a programmable command queue to issue multiple
nand flash commands together.

Signed-off-by: Jan Glauber <jglauber@cavium.com>
Signed-off-by: David Daney <david.daney@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06.
Move the driver to the drivers/mtd/nand/raw dir]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>