i2c-gpio: Also set bus numbers from reg property
commit ff71a9dc016c0c6ae5246dfccf8b07d1e686a278 from
https://github.com/raspberrypi/linux.git

I2C busses can be assigned specific bus numbers using aliases in
Device Tree - string properties where the name is the alias and the
value is the path to the node. The current DT parameter mechanism
does not allow property names to be derived from a parameter value
in any way, so it isn't possible to generate unique or matching
aliases for nodes from an overlay that can generate multiple
instances, e.g. i2c-gpio.

Work around this limitation (at least temporarily) by allowing
the i2c adapter number to be initialised from the "reg" property
if present.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>