drm: xlnx: drv: Allow multiple masters on a drm device
commit  1129555f714fb096c676f8040d75c78bfe7445eb from
https://github.com/Xilinx/linux-xlnx.git

This may not be the ideal solution, but is to allow multiple
masters running on a single drm device, when the application
has the admin priviliege. This enables access to master
only resources such as setting drm planes and etc.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>