MLK-17156-2: ARM: dts: update dts for demo audio in A7 domain
commit  4033f8b63b74fe416bd687256dee477a7edc9ac8 from
https://source.codeaurora.org/external/imx/linux-imx.git

This is dts is for demo SAI + codec in A7 domain, which need to
do i2c hardware rework, connect the wm8960 to i2c7.
for this is for demo usage, so don't need the headphone plug/unplug
event, so remove the iomux for this case.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>