MLK-17631-2 usb: cdns3: gadget: fix the bug the non-ep0 can't work after disconnection
commit  997ccf4804751a6dbef36d5714b1bf5b60c9c08e from
https://source.codeaurora.org/external/imx/linux-imx.git

We add endpoints to ep_match_list when adding gadget module, but
we delete the endpoints from the ep_match_list before set configuration.
When the re-enumeration after the new connection, the ep_match_list
is empty, in that case, the non-ep0s have not configurated, the
transfer on them will be failed.

In this commit, we only delete the endpoints from the list when we
remove the gadget module.

Acked-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>