MLK-17648-1: drm: imx: dcss: add HDR10 module tables
commit  6484afc00e798dee48d933ecf29186bd28d788b6 from
https://source.codeaurora.org/external/imx/linux-imx.git

This commit adds HDR10 tables as a header. Using a FW file is
problematic since the tables need to be available immediately after
boot. After the rootfs is mounted, as is the case for loading a FW file,
it's already too late if some conversion tables are needed.

This usually happens if the output pipe is configured as YUV420.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>