MLK-17648-2: drm: imx: dcss: Load the HDR10 from header file
commit  5cabd42b345565ff06fc888d641219e3b231720e from
https://source.codeaurora.org/external/imx/linux-imx.git

This commit allows one to select if a firmware file is used, for loading
the HDR10 tables, or a header. By default, this will be header file.
This is until a proper way of passing the file from bootloader is found.

Also, fix a minor bug which made parsing the tables over the actual data
limit.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>