MLK-17938: pxp: pxp will hang when output height is 8
commit  bcf745002f1efd0446ff6d18116baf5ec0e4cb30 from
https://source.codeaurora.org/external/imx/linux-imx.git

When output height is eight, yscale of ps engine will be
over two, it not support by pxp.

According to pxp doc, a factor greater than 2 is not
supported with the bilinear filter. so correct it when
this case happen in order to ensure that pxp will be
not hang up.

Reviewed-by: Robby.cai <robby.cai@nxp.com>
Signed-off-by: Guoniu.Zhou <guoniu.zhou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>