MLK-17907-1 soc: imx8: Add compatible to let domu know the errata
commit  7448977b4de3b0ed3be88494985caa80a513f6e9 from
https://source.codeaurora.org/external/imx/linux-imx.git

Because of the TO1.0 issue on i.MX8QM, we need to let DomU
know that errata, then DomU big.Little guest could work correctly.

This is not that good to add compatible for xen vm machine in
imx8 soc code.  Another method is to modify xen toolstack to create
the imx8qm compatible in domu dts, but this is not welcomed, from
xen maitainer's view "Domu should be not binded with a dedicated SoC".

For supporting i.MX8QM, Let's first use the compatible. This might
cause qxp Domu thinking it is 8QM, but we do not support passthrough
on 8QXP, because no smmu.

So it is ok to add the compatible here.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>