pcie: armada8k: add support for comphy initialization
On armada8k platforms the comphy (common phy) must be configured to work
in appropriate mode. Hitherto comphy configuration for pcie was done by
bootloader. As a comphy driver was extended to support pcie comphy
configuration, this patch uses it in pcie driver. Thanks to that there
is no need to relay on the bootloader configuration which could
configured the comphy to different mode.

Change-Id: Ibc78b0411898fd908f95ed6d524bc3e0c9de3bdc
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>