MLK-17747: dsp: use the name of dsp instead of hifi
commit  a8393121b0cbeebad8c01404569da1370e3c6045 from
https://source.codeaurora.org/external/imx/linux-imx.git

In order to avoid the name problem going forward with
integration with Qcom, Qcom has their own dsp and hifi
is competitor, so the hifi name should not be used in
our code.

So use the name of dsp instead of hifi to fix this
problem.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>